+++ CHANGELOG.md
@@ -8,8 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-
 - Add new implementation and view for see who created such services. [#136](https://github.com/DevAlexandreCR/gorda-admin/pull/136)
+
+### Fixed
+- Error when editing a user. [#137](https://github.com/DevAlexandreCR/gorda-admin/pull/137)
 
 # Release Notes for 1.4.x